message

+++ hosts/default/scripts/swipe_up.nix
@@ -5,6 +5,13 @@ pkgs.writeShellScriptBin "swipe_up.sh" ''
 if ! pidof nwg-dock-hyprland; then
     launch_dock.sh &
 fi
+if [[ "$(cat ~/nixos/hosts/default/scripts/DOCK_STATUS.txt)" == "open" ]]; then
+    toggle_hyprpanel_visibility.sh hide
+fi
+if [[ "$(hyprpanel iwv dashboardmenu)" == "true" ]]; then
+    hyprpanel t dashboardmenu
+    exit
+fi
 pkill -f -36 nwg-dock-hyprland
 echo "open" > ~/nixos/hosts/default/scripts/DOCK_STATUS.txt
 ''
